Move Right-click handling into UserInputHandler.
Not sure how I missed this the first try, but there's actually a
onDown() method for GestureDetector. What we can do then is override it,
and then call on rightClick if it's a mouse & a secondary click. This
get rid of a lot of hacky stuff we did.

We still have to keep mAteRightClick so we won't do singleTapUp(), which
is logic for single finger press. Also, we need to catch for right click
still in interceptEvent, so that it doesn't go down into its child views
(Who will try to generate a context menu themselves, and ... things stop
working.)

This also fixes the bug listed below, so now all menus are activated via
ACTION_DOWN only.

Bug: 29548676
Change-Id: I4d4acb3fc1f063379b508bc895c9cf70a2eeb245