On-demand fetch for DocumentInfo when right-click on a root.
RootInfo has information regarding the whole root, but not scoped
narrowly enough for just the top-level document. We need the top level
DocumentInfo to know specifics such as is it writable. This does a fetch
of that when right clicking on a root, and defaults to a timeout of 1000
ms if it takes too long to answer back for whatever reason (Provider
went bad/non-responding, etc).

Bug: 31662523
Change-Id: I2278ea0171fb63a5783f9be619732c857558a1b6