Fix progress bar when loading directories.
Previously, the progress bar was shared between multiple directory
fragments. As a result, when changing directories two instances were
changing visibility of the progress bar depending on a different model,
resulting on prematurely hidden progress bar.

This CL fixes it by moving the progress bar to the directory fragment,
so it's not shared anymore.

Test: Tested manually with a large archive file.
Bug: 32647885
Change-Id: I8ecf459020583f496899ea5465f8bc48386f155c