Add basic support for extracting archives.
This is the first step. It basically replaces the "Copy to" with
"Extract to". Since it's a primary action it's shown as an icon, instead
of the "..." menu.

The next step will be "Extract all", without a need to select all the
files first. It will be also more discoverable.

Finally, this is an experimental UI. It may change.

Test: Unit tests.
Bug: 33641598
Change-Id: I0cd69c351f32b031cd968340a2e7a0079cea9325
(cherry picked from commit 61216ccc96d95d29a65fa8ff68909b7e13a07df2)
(cherry picked from commit 188f2d3cdbf141d14eecca7695a2a1cad715bed7)