Add progress dialog for transfer operations.
Previously we'd show operations progress in the notification,
which is inconvenient and not exposed to users when the user
stays in the app.

This CL solves this issue by showing a progress bar in the
current activity. Note, that the operation will continue when
the user closes the activity. Also, the user can click on
"Continue in background" to continue working on something else
while copying/moving/compressing/etc.

Finally, it's easier for user to cancel an operation now.

Test: Tested manually.
Bug: 35072376
Change-Id: I15b4842270f12fd746adb14dd5329f305ff371e3