Fix NPE caused by not getting a Root DocumentInfo back from task.
We do a force callback on GetRootsDocumentInfoTask, so if the doc we get
back is null, we need to safe-guard against it. It's not very common
that we get no DocumentInfo back (we have a timeout, and most
well-behaving Providers should give us a DocumentInfo), but it is
possible that Provider took too long to respond.

Bug: 35399529
Change-Id: Ic2153c78019c835ea509397d9b208913eb28f7a0