Fix copying files from archives with a keyboard shortcut.
This stopped working, and we started requiring to acquire archives
before performing operations on them. In case of Job.java, we were
resolving the Uris before acquiring, which would fail if the archive
wasn't acquired (in this case users left the archive directory to get
to the destination).

This CL resolves this issue by acquiring archives before constructing
DocumentInfo.

Test: Tested manually by trying to CTRL+C and CTRL+V from archives
      to Downloads.
Bug: 35852761

Change-Id: I9a36d74f37dcf9001a1e3e01ce42aa90be29c806