Fix tests that were broken when running on ARC++ devices.
AuthUI test does not apply to ARC++ devices (The API isn't there), so it
is now an no-op.
RecentsLoaderTest calls on the underlying new query method introduced in
O if feature is on; TestFeatures had everything on by default. Passing
mEnv.features will now enable/disable features according to the device
the tests are running on.

Bug: 37962414
Bug: 37358647
Change-Id: If996c138e74ae0cdb0c8c40ab3eccd5e764bb33b