Fix tests broken by switch to javac compiler.
There was some non-standard business relating to asserts. The switch
to javac wasn't able to emulate the behavior of jack so we needed to update
our code to use Preconditions and for our tests to check for those.
In the process we uncovered (and fixed) a broken test pattern in DragStartListenerTests.

Bug: 64312632
Test: Updated and passing w/ both jack and javac.
Change-Id: Icdccc2bc8eeb340f45b763b9f8745f0479fe0dcb