Extrac DocsUI selection manager.
DocsUI selection manager retains support for being reset.
This accommodation is in support of DocsUI's odd arrangement where
the SelectionManager is created at the activity level, but
shared with fragment instances each corresponding to the viewing of a single directory.
DefaultSelectionManager drops support for being reset. Instances are
now created in a fully valid state.

Bug: 64847011
Test: Passing.
Change-Id: Idaf61b2ec86cb699e90bcee25641f0e288c4e522