Make DocumentUI tests run with R8 enabled
The onFocusItem method was only used from the test. In the app itself it
was always overwritten, so R8 made it abstract. However that  broke the test
failing with an AbstractMethodError.

This fixes the issue with a method which is only referenced from then test
in the same was as was used before:

https://googleplex-android-review.git.corp.google.com/c/platform/packages/apps/DocumentsUI/+/3305625
https://googleplex-android-review.git.corp.google.com/c/platform/packages/apps/DocumentsUI/+/3293384

Test: atest com.android.documentsui.dirlist.KeyInputHandlerTest
Bug: 74746744
Change-Id: Idd91cfb17c90055399fcba601884e441504e85be