Fix duplicate permission info on Adoptable storage
If scoped directory access on Adoptable storage is granted, the granted
info is stored in /data/system/urigrants.xml without uuid info because
it's primary storage. On the other hand, the denial info is stored in
/data/data/com.android.documentsui/shared_prefs/com.android.documentsui_preferences.xml
with uuid info.

This CL fixes not to add uuid info for primary storage when storing the
denial info to shared preference.

Bug: 112321396
Test: manual
Test: atest CtsAppSecurityHostTestCases:ScopedDirectoryAccessTest
Change-Id: I5a8f362895d8b4f1f5f65627cecd738867848d5b