Fix the other app's scoped directory permissions are revoked
When a granted permissions is revoked in Scoped Directory Settings, all
the other app's granted permissions of the same directory are removed.

The cause is that package name is not specified when calling
revokeUriPermission() from ScopedAccessProvider.

This CL fixes to remove only the target application's permission info
from /data/system/urigrants.xml when user revoked a permission in
Settings.

Bug: 112395828
Test: manual
Test: atest CtsAppSecurityHostTestCases:ScopedDirectoryAccessTest
Change-Id: I216ec2f9beea80301911d7b81624b149db4c51a4