Fix using StrictJarFile that is @hide API
In order to do the unbundle DocumentsUI, it is not appropriate to
use the @hide API. StrictJarFile is @hide API.

StrictJarFile has the 4 gigabytes limitation and common compress
support the archive size above 4 gigabytes

ReadableArchiveTest has the implicit assumption that the sequence
order of rows has the consistentence for using the other archive
library. To remove the assumption and move to JUnit 4 style.

fix some checkstyle warning.

Fixes: 112696623
Fixes: 110868242
Test: atest DocumentsUITests

Change-Id: Ibaf0b2cc80cbd322c09791f2287734022572f1e4