Add feature to support multiple archive
To implement ArchiveHandle, ArchiveEntryInputStream and
ArchiveRegistry to support multiple archive type.

ArchiveHandle is a type of common handler to handle the things just
like ZipFile. To get the inputStream for Proxy and get all of
entries in the archive.

ArchiveEntryInputStream is a type of common input stream comes from
supported archive file. And, different archive file has different
method to generate the input stream.

ArchiveRegistry is to maintain the support list and keep the related
generators.

Here are the supported list: 7z, zip, tar, tgz.

To add Apache2 license in DocumentsUI module.

Fixes: 117912708
Test: atest DocumentsUITests

Change-Id: I4eb91af59b1a120a3df17f0546cd612200eea2d2