Fix crash bug while opening zip on recent root
In previous implement, it will still show recents root when open
archive container, and crash when open the zip file again because
its already push docs info to stack and checkArgument will fail.
Refactor the logic isRecents in DocumentStack to fix this behavior.
We also need to add fake DocumentInfo when switch to recent root for
keeping same stack size on root top level. The size on root top is
0 but 1 on other roots, this will cause tool bar issue when open
archive container.

Fix: 127231252
Test: atest DocumentsUIGoogleTests
Change-Id: I49db08b94e28b49b7ef155611cedaae14934a5cf