Inline calls to MimeUtils.
A separate CL in this topic adds MimeMap to the @CorePlatformApi and
deletes MimeUtils. This CL inlines calls to MimeUtils to call MimeMap
instead.

Internally, FileTypeMap was refactored in such a way that this change
is not necessary on branches that have the internal change; therefore,
this CL carries a Merged-In line to skip merging into such branches.

Bug: 120171727
Test: Treehugger

Change-Id: If698d9afad4813ec1c21d511971216a95de9cba3
Merged-In: If62ee1a628feb909cd7b30ecf6e61f6845112d82