Ensure user can only share documents from allowed users.
This is done by placing checking on various levels.

- On the UI level, we check in the loader (DirectoryLoader/Recents Loader).
- We also block querying document in DocumentsAccess
- And we will not return the uris if we are not allowed to access the target
user uri.

Need to follow up on some issues:
Breadcrumb bar disappeared in disabled user
(because we cannot load root document and push to the stack)
Cannot search if selected user is disabled (because of empty stack.
We cannot restart loader in loadDocumentsForCurrentStack)

Bug: 149818298
Bug: 148270816
Test: atest DocumentsUIGoogleTests
Test: manual

Change-Id: I000813d3e66a8c376dbfbd7e4085f403c2e5e0f6