Enable DirectoryFragment.onRefresh works when there is no root doc.
This enables us to reload the directory correctly. On next CL,
we will refresh directory when we listened to some event. (e.g. work
profile is unlocked)

Also now we can perform search in directory loader on a disabled user.

* In DirectoryFragment, refresh stack without root doc
DirectoryFragment will now try to reload rootDoc and push it to stack
onRefresh. This is useful when the root doc was temporarily unable to
load (e.g. work profile was turned off)

* Search across profile on an empty stack
Directory takes care of empty stack now. We will restart loader in
loadDocumentsForCurrentStack. If there is only one queriable user,
the same exception message will be updated.

* ProfileTabs
The tab layout will be updated if the current root does not match.
This could happen when opening a search folder from the other user.

* QuickViewIntentBuilder
Now use a correct packageManager to test intent

Bug: 148270816
Bug: 150600030
Bug: 150799134
Test: atest DocumentsUIGoogleTests
Test: manual

Change-Id: Ib6684e4dd2a257f92ee3784297b5568cbd3d21b2