Add error handling and UI on WP quiet mode / no permission to share
Listen to WP on/off
When broadcast is received, update roots and then refresh the dir if needed.

If canShareAcrossProfile value changes in RootsFragment, refresh
the directory.

Add 2 new inflated message error screens
When WP is in quiet mode, show a button to turn on WP
MODIFY_QUIET_MODE(privileged) is required.

Bug: 148270816
Test: atest DocumentsUIGoogleTests
Test: manual - turning on/off WP > message shown/disappear
Test: manual - changing dpc policy > message shown/disappear
Change-Id: Icd307e503294aae4a8c7a9c3091facee7f6ec814