Show breadcrumb on error state screens
Fixes a bug where the UI jumps when selecting a tab showing an error
state screen. Did not occur on the 'recents' root. The UI jumped because
the breadcrumb disappeared since there is no state on the stack.

We want to maintain this situation where there is no DocumentInfo
objects on the stack in this case, but we still want to show the
breadcrumb. In fact, we don't even use the DocumentInfo on the stack on
the first level of the breadcrumb anyway, so we can actually just update
the adapter to have 1 item in this case and it works.

Fixes: 151097996
Test: manual; this class has no tests currently
Change-Id: Iac06a3b4406367089540a84ee1df8bdf46b201c1