Fix DocsUI cross-profile error screens in landscape
Previously, a fixed height was being used. This worked well on Pixel in
portrait but not on other device types with smaller screens or in
landscape. This CL changes to use either 100 dp or padding on
phones/tablets in portrait, and 0 dp of padding on phones on landscape.
See screenshots in CL comments.

Also adds 48dp of padding to each side of the error message screens as
requested by UX.

Fixes: 152611582
Fixes: 152591331
Test: manual
Change-Id: Ie3ba186191ed1c4d15945a02ce545617ce8de3da