Fix scrolling bug with directory header when loading new tab
There is a bug with the directory header. When you load a new tab, if
you start your vertical scroll from within the header (e.g. breadcrumb,
filters, tabs, app row, etc), then often it simply won't scroll. It
always scrolls if you start from the content window below.

This is almost certainly just a bug with the scrolling mechanism
(it is a CoordinatorLayout > DrawerLayout > CoordinatorLayout >
AppBarLayout > CollapsingToolbarLayout > LinearLayout). These kinds of
bugs are not uncommon in Android with complex nested views and it's a
difficult problem to search for.

Wrapping the problem view (the directory header) in a NestedScrollView
fixed it. See videos in comments.

This CL is dependent on the assumption of rolling back
I657c404928d80759f4de5690690a5dd2196be8e8, which broke scrolling.

Fixes: 151145028
Test: manual
Change-Id: I14bca8c7649558d114fbd607507bd27efff30a2d