Fix skipping the last address line and avoid appending nulls.
getMaxAddressLineIndex() returns the largest index instead of the
number of lines, so a less-than-or-equal check is the correct
condition here, or we'll lose the last line. Address lines may also be
null, and in case of null having an empty line is still better than
having a line saying "null".

Fixes: 155585322
Test: presubmit
Change-Id: I150db3ced2fbcdf73faf0cc6adf618b13d5acfca