DocumentsUI: Declare intents for App enumeration
With the introduction of App enumeration in Android R, from the
FilesApp, a file can no longer be opened by an andriod app. The
Android app cannot be started. This change is to fix this issue
by declaring intents in DocumentsUI's manifest so that it is
allowed to interact with other apps.

Bug: 152288750
Test: Select a file in the FilesApp. There should be an option to open
the file with an android app and the app can be started.

Change-Id: Ia3361e6e5abf50f1fe7c80c7b00d0ef93916d875