Fix preview icon tapping behavior when magnifying the screen
We user the unscaled coordinate to determine whether the event is
in the bounds of the view. It doesn't work in magnification scenario
because the view size is not changed and getRawX()/getRawY() is the actual
position on the screen.

To fix it, we determine it in the parent's coordinate. In our case
the event is dispatched from RecyclerView, so the calculation is in
Recyclerview's coordinate

Bug: 152265995
Test: manual test
Change-Id: If6e2497641d16b92542f3fc6576929c52fe48fbb