Remove @SortDimensionId
The sort dimension ID can't be an @IntDef because the values are not
constant when DocumentsUI is compiled as a library so that it can be
compiled once for DocumentsUI and DocumentsUIGoogle.

This relands Ie6ddc3869c6f758b75a048a2e5fe2e0dc9305424.

Bug: 294256649
Test: m DocumentsUI
Merged-In: I5f7c7c2ac90fb852e7de47299b159f2fe71ca76a
Change-Id: I5f7c7c2ac90fb852e7de47299b159f2fe71ca76a