Block taking recents screenshot if private profile available
DocsUI shows separate tab when private profile is on.
User can move DocUI to the background and then lock private
profile. Recents preview screen shows screenshot
of the DocUI at the last moment it was in foreground
(when the space was unlocked). As a result Recents preview
can leak whether private profile exists on device.
This change sets setting to show blank
screen in the recents when private profile is avaialble.

Bug: 329241706
Test: manual, https://b.corp.google.com/issues/329241706#comment3
Change-Id: I05914391db7bd15a88c6043d14265682ffd8437c